Change structure to match new restructure of PR28
