--- docs/source/test_file_ops.rst
@@ -1,64 +1,0 @@
-
-test_file_ops
-===============
-**test_file_ops** contains tests for file_ops functions
-
-.. autofunction:: test_file_ops.test_output_path
-
-.. autofunction:: test_file_ops.test_convert_str_to_pathlib_path
-
-.. autofunction:: test_file_ops.test_path_exists
-
-.. autofunction:: test_file_ops.test_extract_zip_to_output_path
-
-.. autofunction:: test_file_ops.test_extract_zip_no_output_path
-
-.. autofunction:: test_file_ops.test_zip_not_deleted
-
-.. autofunction:: test_file_ops.test_get_bucket_filenames_s3
-
-.. autofunction:: test_upload_obj_dataframe_csv
-
-.. autofunction:: test_upload_obj_dataframe_parquet
-
-.. autofunction:: test_upload_obj_dict_json
-
-.. autofunction:: test_upload_obj_list_csv
-
-.. autofunction:: test_upload_obj_list_txt
-
-.. autofunction:: test_upload_obj_list_json
-
-.. autofunction:: test_upload_obj_str_txt
-
-.. autofunction:: test_upload_obj_array_csv
-
-.. autofunction:: test_upload_obj_array_parquet
-
-.. autofunction:: test_upload_obj_unsup_data
-
-.. autofunction:: test_dowload_obj_dataframe_csv
-
-.. autofunction:: test_dowload_obj_dataframe_parquet
-
-.. autofunction:: test_dowload_obj_dict_json
-
-.. autofunction:: test_dowload_obj_list_csv
-
-.. autofunction:: test_dowload_obj_list_txt
-
-.. autofunction:: test_dowload_obj_list_json
-
-.. autofunction:: test_dowload_obj_str_txt
-
-.. autofunction:: test_dowload_obj_array_csv
-
-.. autofunction:: test_dowload_obj_array_parquet
-
-.. autofunction:: test_dowload_obj_unsup_data
-
-.. autofunction:: test_dowload_obj_exeption
-
-.. autofunction:: test_file_ops.test_download_file
-
-.. autofunction:: test_file_ops.test_upload_file
