+++ tests/loading_saving/test_S3.py
@@ -1,7 +1,5 @@
 import shutil
-import os
 from pathlib import Path
-import sys
 import pytest
 import pandas as pd
 import numpy as np
@@ -9,9 +7,7 @@ from typing import List
 import boto3
 from moto import mock_s3
 import io
-
-sys.path.insert(1, "nesta_ds_utils/")
-import file_ops
+from nesta_ds_utils.loading_saving import S3
 
 TEST_DATAFRAME = pd.DataFrame({"test": [0, 0]})
 TEST_DICT = {"test": [0, 0]}
@@ -20,85 +16,14 @@ TEST_STR = "test"
 TEST_NP_ARRAY = np.array([0, 0])
 
 
-@pytest.fixture
-def test_output_path():
-    """Generates pathlib.Path to dump intermediate test data.
-
-    Yields:
-        Output path to dump intermetiate test data.
-    """
-    test_output_path = file_ops._convert_str_to_pathlib_path("tests/temp/")
-    file_ops.make_path_if_not_exist(test_output_path)
-    yield test_output_path
-    shutil.rmtree("tests/temp/")
-
-
-def test_convert_str_to_pathlib_path(test_output_path: Path):
-    """Tests that file_ops method convert_str_to_pathlib_path.
-        returns type patlib.Path.
-
-    Args:
-        test_output_path (pathlib.Path): output path to dump intermetiate test data
-    """
-    assert isinstance(test_output_path, Path)
-
-
-def test_path_exists(test_output_path: Path):
-    """Tests that the path generated by file_ops method
-        make_path_if_not_exist exists.
-
-    Args:
-        test_output_path (pathlib.Path): Output path to dump intermetiate test data.
-    """
-    assert test_output_path.exists()
-
-
-def test_extract_zip_to_output_path(test_output_path: Path):
-    """Tests that the file_ops method extractall dumped a
-        text file to an output path and that text file could be read.
-
-    Args:
-        test_output_path (pathlib.Path): Output path to dump intermetiate test data.
-    """
-    file_ops.extractall(
-        "tests/artifacts/dummy_zip.zip", test_output_path, delete_zip=False
-    )
-
-    with open("tests/temp/dummy_text_in_zip.txt", "r") as f:
-        text = f.read()
-        assert text == "'Hello World'"
-
-
-def test_extract_zip_no_output_path():
-    """Tests that the file_ops method extractall dumped a
-    text file to the directory of the zip file.
-    """
-    file_ops.extractall("tests/artifacts/dummy_zip.zip", delete_zip=False)
-
-    with open("tests/artifacts/dummy_text_in_zip.txt", "r") as f:
-        text = f.read()
-        assert text == "'Hello World'"
-
-    os.remove("tests/artifacts/dummy_text_in_zip.txt")
-
-
-def test_zip_not_deleted():
-    """Tests that the zip file was not deleted when extractall
-    was called without a delete_zip flag.
-    """
-    assert file_ops._convert_str_to_pathlib_path(
-        "tests/artifacts/dummy_zip.zip"
-    ).exists()
-
-
 @mock_s3
 def test_get_bucket_filenames_s3():
     """Test that get_bucket_filenames_s3 returns a list containing dummy.csv."""
     conn = boto3.resource("s3", region_name="us-east-1")
     conn.create_bucket(Bucket="mybucket")
-    s3 = boto3.client("s3")
-    s3.upload_fileobj(io.BytesIO(b"Test"), "mybucket", "dummy.csv")
-    assert file_ops.get_bucket_filenames_s3("mybucket", "") == ["dummy.csv"]
+    mock_client = boto3.client("s3")
+    mock_client.upload_fileobj(io.BytesIO(b"Test"), "mybucket", "dummy.csv")
+    assert S3.get_bucket_filenames_s3("mybucket", "") == ["dummy.csv"]
 
 
 @mock_s3
@@ -108,8 +33,8 @@ def test_upload_obj_dataframe_csv():
     """
     conn = boto3.resource("s3", region_name="us-east-1")
     conn.create_bucket(Bucket="test-bucket")
-    s3 = boto3.client("s3")
-    file_ops.upload_obj(TEST_DATAFRAME, "test-bucket", "dummy.csv")
+    mock_client = boto3.client("s3")
+    S3.upload_obj(TEST_DATAFRAME, "test-bucket", "dummy.csv")
 
 
 @mock_s3
@@ -119,8 +44,8 @@ def test_upload_obj_dataframe_parquet():
     """
     conn = boto3.resource("s3", region_name="us-east-1")
     conn.create_bucket(Bucket="test-bucket")
-    s3 = boto3.client("s3")
-    file_ops.upload_obj(TEST_DATAFRAME, "test-bucket", "dummy.parquet")
+    mock_client = boto3.client("s3")
+    S3.upload_obj(TEST_DATAFRAME, "test-bucket", "dummy.parquet")
 
 
 @mock_s3
@@ -128,8 +53,8 @@ def test_upload_obj_dict_json():
     """Tests that upload_obj does not return an exeption."""
     conn = boto3.resource("s3", region_name="us-east-1")
     conn.create_bucket(Bucket="test-bucket")
-    s3 = boto3.client("s3")
-    file_ops.upload_obj(TEST_DICT, "test-bucket", "dummy.json")
+    mock_client = boto3.client("s3")
+    S3.upload_obj(TEST_DICT, "test-bucket", "dummy.json")
 
 
 @mock_s3
@@ -139,8 +64,8 @@ def test_upload_obj_list_csv():
     """
     conn = boto3.resource("s3", region_name="us-east-1")
     conn.create_bucket(Bucket="test-bucket")
-    s3 = boto3.client("s3")
-    file_ops.upload_obj(TEST_LIST, "test-bucket", "dummy.csv")
+    mock_client = boto3.client("s3")
+    S3.upload_obj(TEST_LIST, "test-bucket", "dummy.csv")
 
 
 @mock_s3
@@ -150,8 +75,8 @@ def test_upload_obj_list_txt():
     """
     conn = boto3.resource("s3", region_name="us-east-1")
     conn.create_bucket(Bucket="test-bucket")
-    s3 = boto3.client("s3")
-    file_ops.upload_obj(TEST_LIST, "test-bucket", "dummy.txt")
+    mock_client = boto3.client("s3")
+    S3.upload_obj(TEST_LIST, "test-bucket", "dummy.txt")
 
 
 @mock_s3
@@ -161,8 +86,8 @@ def test_upload_obj_list_json():
     """
     conn = boto3.resource("s3", region_name="us-east-1")
     conn.create_bucket(Bucket="test-bucket")
-    s3 = boto3.client("s3")
-    file_ops.upload_obj([TEST_DICT, TEST_DICT], "test-bucket", "dummy.json")
+    mock_client = boto3.client("s3")
+    S3.upload_obj([TEST_DICT, TEST_DICT], "test-bucket", "dummy.json")
 
 
 @mock_s3
@@ -172,8 +97,8 @@ def test_upload_obj_str_txt():
     """
     conn = boto3.resource("s3", region_name="us-east-1")
     conn.create_bucket(Bucket="test-bucket")
-    s3 = boto3.client("s3")
-    file_ops.upload_obj(TEST_STR, "test-bucket", "dummy.txt")
+    mock_client = boto3.client("s3")
+    S3.upload_obj(TEST_STR, "test-bucket", "dummy.txt")
 
 
 @mock_s3
@@ -183,8 +108,8 @@ def test_upload_obj_array_csv():
     """
     conn = boto3.resource("s3", region_name="us-east-1")
     conn.create_bucket(Bucket="test-bucket")
-    s3 = boto3.client("s3")
-    file_ops.upload_obj(TEST_NP_ARRAY, "test-bucket", "dummy.csv")
+    mock_client = boto3.client("s3")
+    S3.upload_obj(TEST_NP_ARRAY, "test-bucket", "dummy.csv")
 
 
 @mock_s3
@@ -194,8 +119,8 @@ def test_upload_obj_array_parquet():
     """
     conn = boto3.resource("s3", region_name="us-east-1")
     conn.create_bucket(Bucket="test-bucket")
-    s3 = boto3.client("s3")
-    file_ops.upload_obj(TEST_NP_ARRAY, "test-bucket", "dummy.parquet")
+    mock_client = boto3.client("s3")
+    S3.upload_obj(TEST_NP_ARRAY, "test-bucket", "dummy.parquet")
 
 
 @mock_s3
@@ -205,8 +130,8 @@ def test_upload_obj_unsup_data():
     """
     conn = boto3.resource("s3", region_name="us-east-1")
     conn.create_bucket(Bucket="test-bucket")
-    s3 = boto3.client("s3")
-    file_ops.upload_obj(0, "test-bucket", "dummy.pkl")
+    mock_client = boto3.client("s3")
+    S3.upload_obj(0, "test-bucket", "dummy.pkl")
 
 
 @mock_s3
@@ -216,12 +141,12 @@ def test_dowload_obj_dataframe_csv():
     """
     conn = boto3.resource("s3", region_name="us-east-1")
     conn.create_bucket(Bucket="test-bucket")
-    s3 = boto3.client("s3")
-    s3.upload_file("tests/artifacts/dummy_dataframe.csv", "test-bucket", "dummy.csv")
+    mock_client = boto3.client("s3")
+    mock_client.upload_file(
+        "tests/artifacts/dummy_dataframe.csv", "test-bucket", "dummy.csv"
+    )
     assert (
-        file_ops.download_obj("test-bucket", "dummy.csv", download_as="dataframe").test[
-            0
-        ]
+        S3.download_obj("test-bucket", "dummy.csv", download_as="dataframe").test[0]
         == 0
     )
 
@@ -233,14 +158,12 @@ def test_dowload_obj_dataframe_parquet():
     """
     conn = boto3.resource("s3", region_name="us-east-1")
     conn.create_bucket(Bucket="test-bucket")
-    s3 = boto3.client("s3")
-    s3.upload_file(
+    mock_client = boto3.client("s3")
+    mock_client.upload_file(
         "tests/artifacts/dummy_dataframe.parquet", "test-bucket", "dummy.parquet"
     )
     assert (
-        file_ops.download_obj(
-            "test-bucket", "dummy.parquet", download_as="dataframe"
-        ).test[0]
+        S3.download_obj("test-bucket", "dummy.parquet", download_as="dataframe").test[0]
         == 0
     )
 
@@ -252,13 +175,12 @@ def test_dowload_obj_dict_json():
     """
     conn = boto3.resource("s3", region_name="us-east-1")
     conn.create_bucket(Bucket="test-bucket")
-    s3 = boto3.client("s3")
-    s3.upload_file("tests/artifacts/dummy_dict.json", "test-bucket", "dummy.json")
+    mock_client = boto3.client("s3")
+    mock_client.upload_file(
+        "tests/artifacts/dummy_dict.json", "test-bucket", "dummy.json"
+    )
     assert (
-        file_ops.download_obj("test-bucket", "dummy.json", download_as="dict")["test"][
-            0
-        ]
-        == 0
+        S3.download_obj("test-bucket", "dummy.json", download_as="dict")["test"][0] == 0
     )
 
 
@@ -269,11 +191,11 @@ def test_dowload_obj_list_csv():
     """
     conn = boto3.resource("s3", region_name="us-east-1")
     conn.create_bucket(Bucket="test-bucket")
-    s3 = boto3.client("s3")
-    s3.upload_file("tests/artifacts/dummy_list.csv", "test-bucket", "dummy.csv")
-    assert (
-        file_ops.download_obj("test-bucket", "dummy.csv", download_as="list")[0] == "0"
+    mock_client = boto3.client("s3")
+    mock_client.upload_file(
+        "tests/artifacts/dummy_list.csv", "test-bucket", "dummy.csv"
     )
+    assert S3.download_obj("test-bucket", "dummy.csv", download_as="list")[0] == "0"
 
 
 @mock_s3
@@ -283,11 +205,11 @@ def test_dowload_obj_list_txt():
     """
     conn = boto3.resource("s3", region_name="us-east-1")
     conn.create_bucket(Bucket="test-bucket")
-    s3 = boto3.client("s3")
-    s3.upload_file("tests/artifacts/dummy_list.txt", "test-bucket", "dummy.txt")
-    assert (
-        file_ops.download_obj("test-bucket", "dummy.txt", download_as="list")[0] == "0"
+    mock_client = boto3.client("s3")
+    mock_client.upload_file(
+        "tests/artifacts/dummy_list.txt", "test-bucket", "dummy.txt"
     )
+    assert S3.download_obj("test-bucket", "dummy.txt", download_as="list")[0] == "0"
 
 
 @mock_s3
@@ -297,12 +219,12 @@ def test_dowload_obj_list_json():
     """
     conn = boto3.resource("s3", region_name="us-east-1")
     conn.create_bucket(Bucket="test-bucket")
-    s3 = boto3.client("s3")
-    s3.upload_file("tests/artifacts/dummy_list.json", "test-bucket", "dummy.json")
+    mock_client = boto3.client("s3")
+    mock_client.upload_file(
+        "tests/artifacts/dummy_list.json", "test-bucket", "dummy.json"
+    )
     assert (
-        file_ops.download_obj("test-bucket", "dummy.json", download_as="list")[0][
-            "test"
-        ][0]
+        S3.download_obj("test-bucket", "dummy.json", download_as="list")[0]["test"][0]
         == 0
     )
 
@@ -314,11 +236,9 @@ def test_dowload_obj_str_txt():
     """
     conn = boto3.resource("s3", region_name="us-east-1")
     conn.create_bucket(Bucket="test-bucket")
-    s3 = boto3.client("s3")
-    s3.upload_file("tests/artifacts/dummy_str.txt", "test-bucket", "dummy.txt")
-    assert (
-        file_ops.download_obj("test-bucket", "dummy.txt", download_as="str") == "test\n"
-    )
+    mock_client = boto3.client("s3")
+    mock_client.upload_file("tests/artifacts/dummy_str.txt", "test-bucket", "dummy.txt")
+    assert S3.download_obj("test-bucket", "dummy.txt", download_as="str") == "test\n"
 
 
 @mock_s3
@@ -328,12 +248,11 @@ def test_dowload_obj_array_csv():
     """
     conn = boto3.resource("s3", region_name="us-east-1")
     conn.create_bucket(Bucket="test-bucket")
-    s3 = boto3.client("s3")
-    s3.upload_file("tests/artifacts/dummy_array.csv", "test-bucket", "dummy.csv")
-    assert (
-        file_ops.download_obj("test-bucket", "dummy.csv", download_as="np.array")[0]
-        == 0
+    mock_client = boto3.client("s3")
+    mock_client.upload_file(
+        "tests/artifacts/dummy_array.csv", "test-bucket", "dummy.csv"
     )
+    assert S3.download_obj("test-bucket", "dummy.csv", download_as="np.array")[0] == 0
 
 
 @mock_s3
@@ -343,13 +262,12 @@ def test_dowload_obj_array_parquet():
     """
     conn = boto3.resource("s3", region_name="us-east-1")
     conn.create_bucket(Bucket="test-bucket")
-    s3 = boto3.client("s3")
-    s3.upload_file(
+    mock_client = boto3.client("s3")
+    mock_client.upload_file(
         "tests/artifacts/dummy_array.parquet", "test-bucket", "dummy.parquet"
     )
     assert (
-        file_ops.download_obj("test-bucket", "dummy.parquet", download_as="np.array")[0]
-        == 0
+        S3.download_obj("test-bucket", "dummy.parquet", download_as="np.array")[0] == 0
     )
 
 
@@ -360,9 +278,11 @@ def test_dowload_obj_unsup_data():
     """
     conn = boto3.resource("s3", region_name="us-east-1")
     conn.create_bucket(Bucket="test-bucket")
-    s3 = boto3.client("s3")
-    s3.upload_file("tests/artifacts/dummy_unsupp.pkl", "test-bucket", "dummy.pkl")
-    assert file_ops.download_obj("test-bucket", "dummy.pkl") == 0
+    mock_client = boto3.client("s3")
+    mock_client.upload_file(
+        "tests/artifacts/dummy_unsupp.pkl", "test-bucket", "dummy.pkl"
+    )
+    assert S3.download_obj("test-bucket", "dummy.pkl") == 0
 
 
 @mock_s3
@@ -370,9 +290,9 @@ def test_dowload_obj_exeption():
     """Tests that download_obj returns an exception for unsupported file type."""
     conn = boto3.resource("s3", region_name="us-east-1")
     conn.create_bucket(Bucket="test-bucket")
-    s3 = boto3.client("s3")
+    mock_client = boto3.client("s3")
     with pytest.raises(Exception):
-        file_ops.download_obj("test-bucket", "dummy.html")
+        S3.download_obj("test-bucket", "dummy.html")
 
 
 @mock_s3
@@ -380,9 +300,9 @@ def test_download_file():
     """Tests that download_file download mock txt file."""
     conn = boto3.resource("s3", region_name="us-east-1")
     conn.create_bucket(Bucket="test-bucket")
-    s3 = boto3.client("s3")
-    s3.upload_fileobj(io.BytesIO(b"Test"), "test-bucket", "dummy.csv")
-    file_ops.download_file("dummy.csv", "test-bucket", "tests/temp/dummy.csv")
+    mock_client = boto3.client("s3")
+    mock_client.upload_fileobj(io.BytesIO(b"Test"), "test-bucket", "dummy.csv")
+    S3.download_file("dummy.csv", "test-bucket", "tests/temp/dummy.csv")
     with open("tests/temp/dummy.csv", "r") as f:
         text = f.read()
         assert text == "Test"
@@ -394,10 +314,10 @@ def test_upload_file():
     """Tests that upload_file upload mock txt file."""
     conn = boto3.resource("s3", region_name="us-east-1")
     conn.create_bucket(Bucket="test-bucket")
-    s3 = boto3.client("s3")
-    file_ops.upload_file("tests/artifacts/dummy.txt", "test-bucket", "dummy.csv")
+    mock_client = boto3.client("s3")
+    S3.upload_file("tests/artifacts/dummy.txt", "test-bucket", "dummy.csv")
     Path("tests/temp/").mkdir(parents=True)
-    s3.download_file("test-bucket", "dummy.csv", "tests/temp/dummy.csv")
+    mock_client.download_file("test-bucket", "dummy.csv", "tests/temp/dummy.csv")
     with open("tests/temp/dummy.csv", "r") as f:
         text = f.read()
         assert text == "Test\n"
